renaming `checkout.context`
