+++ src/app/checkout/[show]/checkout-summary-card.tsx
@@ -4,7 +4,7 @@ import { ChangeEvent } from 'react';
 
 import { Card } from '@/components';
 
-import { useCheckoutContext } from '../checkout-context';
+import { useCheckoutContext } from '../checkout.context';
 
 interface CheckoutSummaryCardProps {
   price: number;
