+++ src/app/checkout/layout.tsx
@@ -1,5 +1,5 @@
 import { PropsWithChildren } from 'react';
-import { CheckoutProvider } from './checkout-context';
+import { CheckoutProvider } from './checkout.context';
 
 export default function CheckoutLayout({ children }: PropsWithChildren) {
   return <CheckoutProvider>{children}</CheckoutProvider>;
